Update patch set 1
Patch Set 1:

About bumping the version: I was thinking that since Sai and Ken already know about 2.1.1, which fixes model images not being copied, I shouldn't add more changes to that version number, but should start another one.

For the URLs in CatalogServiceTest, because this is a test, I'd prefer not to make these constants, that is, I'd prefer to have both the input values and the corresponding expected result values appear explicitly, rather than having the test code compute the expected result values.  For ControllerTest, I will use string constants as, in that case, the different values of the string have actual meaning.

I'll post a patch, shortly.

Patch-set: 1