Update patch set 3
Patch Set 3:

(1 comment)

I think I've addressed the concern that this was returning all catalogs,  rather than only visible ones, consistent with the mechanism already used for returning only public solutions/revisions.  This happens around CatalogServiceImpl.java line 102.  This catalog filtering behavior is verified in CatalogServiceTest.testCatalogService() which tests that 2 catalogs are returned even though the mock data in mockCDSPortalCatalogsResponse.json has 3 catalogs (two with a code of PB, and one with an accessTypeCode of PR, which is filtered out).

Patch-set: 3