Update patch set 1
Patch Set 1: Code-Review-1

This is a functionality change but does not change the version, is that acceptable?

I'm voting -1 because of missing release notes, and because I expect a change to be cherry picked to master branch at the same time.

Patch-set: 1
Reviewer: Gerrit User 4 <4@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1