Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> 
> Justin, does this satisfy your concerns?

Yes for Clio - we will get software registered. We need to be able to test this patch out .. but we were not able to get federation service connecting + Parag who was working on setting up federation for Clio in our lab. 

Concerns we can address in Demeter
1. Decoupling code here using kafka/pub-sub
2. How to not allow any use of solution if LUM service is not there or has a failure. 
3. We may need to have a state in database that software was registered with LUM service.

Patch-set: 3