Update patch set 3
Patch Set 3:

(30 comments)

See comments.  A few highlights: The removal of existing signatures on methods in ClientBase unnecessarily breaks backward compatibility of the library, and results in the need for a lot of other otherwise unnecessary changes, as well as breaking code for other users of the client library.  The addition of completely independent controllers for model data breaks the conventions in the existing code and also results in extraneous code duplication.  The GatewayClient needs a new API access method to invoke the new internal API.  The client library needs test code for the new API access methods.  The gateway needs test code for the new APIs.  Changes to existing javadoc on existing code should not be made unless either the changes make the javadoc notably clearer or the behavior of the method is changing.  New code should use indentation and other whitespace conventions in alignment with the existing code.  Whitespace shouldn't be just changed for no good reason in code that is otherwise not being changed.

Patch-set: 3