Update patch set 6
Patch Set 6: Code-Review-1

(22 comments)

See comments.

You still haven't added test cases for the new code.
You still haven't added the corresponding call to the GatewayClient.
You're still making changes to the dependencies in pom.xml files based on an apparent misunderstanding of the differences between "dependencies" and "dependencyManagement" and an incorrect assumption that listing a dependency in "dependencyManagement" creates a dependency, in that pom.xml on the item described by the dependency.
You're still making irrelevant changes to javadoc and white space in a delivery of a "critical piece."
The federation controller should not be making direct calls to the logstash API.  It should be invoking a client, using an instance created in Clients.java, so that the controller and backend code can be independently tested in the unit tests.
The data representation of the call and return parameters to the new APIs should be properly defined, not just "JsonNode" with a text description in the JavaDoc.

Patch-set: 6
Label: Code-Review=-1