Update patch set 7
Patch Set 7:

(4 comments)

> Patch Set 6: Code-Review-1
> 
> (22 comments)
> 
> See comments.
> 
> You still haven't added test cases for the new code.

Did you see FederationControllerTest? There is a test case there and we can add more as required. In patch 7 it has been improved to not call register, unregister, etc

> You still haven't added the corresponding call to the GatewayClient.

I am not following you here? what corresponding call?

> You're still making changes to the dependencies in pom.xml files based on an apparent misunderstanding of the differences between "dependencies" and "dependencyManagement" and an incorrect assumption that listing a dependency in "dependencyManagement" creates a dependency, in that pom.xml on the item described by the dependency.

The problem we were facing is that the build doesn't work consistently and we found if we add the version number it works consistently. I have been replacing that with the property instead of having a duplicate version number to manage. Is this an ok compromise?

> You're still making irrelevant changes to javadoc and white space in a delivery of a "critical piece."

I can remove all javadoc changes that are not relevant.. I think I was just trying to fix up javadoc warnings .. I went a little overboard. Reverting those additional javadoc warning fixes.

> The federation controller should not be making direct calls to the logstash API.  It should be invoking a client, using an instance created in Clients.java, so that the controller and backend code can be independently tested in the unit tests.

Ok I have made an attempt to use the client api and added LogstashClient and LogstashService (and impl)

> The data representation of the call and return parameters to the new APIs should be properly defined, not just "JsonNode" with a text description in the JavaDoc.

We need to have an open ended object since each model will have different parameters.. there is some items that are the pre defined so maybe we can have a mix of defined and not defined at this point. This will require some design changes.  I need to come back on this one.

Patch-set: 7
Label: Code-Review=0