Update patch set 2
Patch Set 2:

"3.1.0.1" is not Semantic Versioning compliant.

I suggest 3.0.4.

However I'm also unclear on why it is necessary to build a special version of Federation with the 3.1.0 CDS client, since the version change of CDS (from 3.0.0 to 3.1.0) indicates that it does not create any incompatibility (is not a breaking change) - it only adds features (which were obviously not needed by Federation).

Patch-set: 2
Reviewer: Gerrit User 288 <288@655300a2-7a56-4504-9153-881edd907bea>
Label: Verified=0